+++ src/Resources/contao/languages/de/tl_vereinsregister.php
@@ -3,159 +3,78 @@
 /**
  * Backend-Modul: Übersetzungen im Eingabeformular
  */
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["name_legend"] = "Vereinsname";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['name_legend'] = 'Vereinsname';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['name'] = array('Name', 'Geben Sie hier den aktuellen Vereinsnamen ein.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['association'] = array('Verband/Bund', 'Verein ist ein Verband/Bund mehrerer Vereine oder hat einen überregionalen Charakter.');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["name"] = array('Name', 'Geben Sie hier den aktuellen Vereinsnamen ein.');
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["association"] = array('Verband/Bund', 'Verein ist ein Verband/Bund mehrerer Vereine oder hat einen überregionalen Charakter.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['place_legend'] = 'Niederlassung';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['continent'] = array('Kontinent', 'Geben Sie hier den Kontinent ein.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['land'] = array('Land', 'Geben Sie hier das Land ein.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['region'] = array('Region', 'Geben Sie hier die Region oder das Bundesland ein.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['city'] = array('Ort', 'Geben Sie hier den Ort ein.');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["place_legend"] = "Niederlassung";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['time_legend'] = 'Gründung und Auflösung';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['timerange'] = array('Zeitraum', 'Von wann bis wann war der Verein aktiv?');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['periodStartDate'] = array('Beginn der Periode', 'Beginndatum des Zeitabschnitts/Eintrags im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['periodEndDate'] = array('Ende der Periode', 'Endedatum des Zeitabschnitts/Eintrags im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['foundingDate'] = array('Gründungsdatum', 'Datum der Gründung im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ. Gilt global inkl. aller Vorgängervereine.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['resolutionDate'] = array('Auflösungsdatum', 'Datum der Auflösung im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ. Kann frei bleiben, wenn es einen Nachfolgeverein gibt.');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["continent"][0] = "Kontinent";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["continent"][1] = "Geben Sie hier den Kontinent ein.";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['before_legend'] = 'Vorgänger-Vereine';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['addBefore'] = array('Vorgänger-Vereine hinzufügen', 'Hier können Sie ein oder mehrere Vorgänger-Vereine hinzufügen.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['beforeClubs'] = array('Bitte Verein(e) auswählen:', 'Bitte wählen Sie ein oder mehrere Vereine aus.');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["land"][0] = "Land";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["land"][1] = "Geben Sie hier das Land ein.";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['separation_legend'] = 'Abgespaltete Vereine';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['addSeparation'] = array('Abgespaltete Vereine hinzufügen', 'Abgespaltete Vereine hinzufügen');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["region"][0] = "Region";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["region"][1] = "Geben Sie hier die Region oder das Bundesland ein.";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['separationClubs'] = 'Vereine';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['separationClub'] = array('Verein', 'Name des abgespalteten Vereins');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['separationDate'] = array('Datum der Trennung', 'Wann spaltete sich dieser Verein ab?');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["city"][0] = "Ort";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["city"][1] = "Geben Sie hier den Ort ein.";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['after_legend'] = 'Nachfolger-Vereine';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['addAfter'] = array('Nachfolger-Vereine hinzufügen', 'Hier können Sie ein oder mehrere Nachfolger-Vereine hinzufügen.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['afterClubs'] = array('Bitte Verein(e) auswählen:', 'Bitte wählen Sie ein oder mehrere Vereine aus.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['afterCause'] = array('Grund', 'Bitte wählen Sie Grund für die Nachfolge aus.');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["time_legend"] = "Gründung und Auflösung";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['info_legend'] = 'Information';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['info'] = array('Vereinsbeschreibung', 'Geben Sie hier eine kurze Beschreibung des Vereins ein.');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["timerange"][0] = "Zeitraum";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["timerange"][1] = "Von wann bis wann war der Verein aktiv?";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['image_legend'] = 'Bild/Logo-Einstellungen';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['addImage'] = array('Ein Bild/Logo hinzufügen', 'Dem Ereignis ein Bild/Logo hinzufügen.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['singleSRC'] = array('Quelldatei', 'Bitte wählen Sie einBild/Logo aus der Dateiübersicht.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['alt'] = array('Alternativer Text', 'Hier können Sie einen alternativen Text für das Bild eingeben (<i>alt</i>-Attribut)');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['title'] = array('Titel', 'Hier können Sie den Titel des Bildes eingeben (<i>title</i>-Attribut)');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['size'] = array('Bildbreite und Bildhöhe', 'Hier können Sie die Abmessungen des Bildes und den Skalierungsmodus festlegen.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['imageUrl'] = array('Bildlink-Adresse', 'Eine eigene Bildlink-Adresse überschreibt den Lightbox-Link, so daß das Bild nicht mehr in der Großansicht dargestellt werden kann.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['imageTitle'] = array('Bildtitel', 'Titel-Tag für das Bild');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['fullsize'] = array('Großansicht/Neues Fenster', 'Großansicht des Bildes in einer Lightbox bzw. den Link in einem neuem Browserfenster öffnen.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['caption'] = array('Bildunterschrift', 'Hier können Sie einen kurzen Text eingeben, der unterhalb des Bildes angezeigt wird.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['floating'] = array('Bildausrichtung', 'Bitte legen Sie fest, wie das Bild ausgerichtet werden soll.');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['overwriteMeta'] = array('Metadaten überschreiben', 'Metadaten überschreiben.');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["periodStartDate"][0] = "Beginn der Periode";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["periodStartDate"][1] = "Beginndatum des Zeitabschnitts/Eintrags im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['link_legend'] = 'Weblink';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['url'] = array('Homepage', 'Homepage des Vereins');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["periodEndDate"][0] = "Ende der Periode";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["periodEndDate"][1] = "Endedatum des Zeitabschnitts/Eintrags im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['live_legend'] = 'Lebensdaten';
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["foundingDate"][0] = "Gründungsdatum";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["foundingDate"][1] = "Datum der Gründung im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ. Gilt global inkl. aller Vorgängervereine.";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['intern_legend'] = 'Intern';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['intern'] = array('Bemerkungen', 'Interne Bemerkungen');
 
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["resolutionDate"][0] = "Auflösungsdatum";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["resolutionDate"][1] = "Datum der Auflösung im Format JJJJ, MM.JJJJ oder TT.MM.JJJJ. Kann frei bleiben, wenn es einen Nachfolgeverein gibt.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["before_legend"] = "Vorgänger-Vereine";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["addBefore"][0] = "Vorgänger-Vereine hinzufügen";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["addBefore"][1] = "Hier können Sie ein oder mehrere Vorgänger-Vereine hinzufügen.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["beforeClubs"][0] = "Bitte Verein(e) auswählen:";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["beforeClubs"][1] = "Bitte wählen Sie ein oder mehrere Vereine aus.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["separation_legend"] = "Abgespaltete Vereine";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["addSeparation"][0] = "Abgespaltete Vereine hinzufügen";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["addSeparation"][1] = "Abgespaltete Vereine hinzufügen";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["separationClubs"] = "Vereine";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["separationClub"][0] = "Verein";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["separationClub"][1] = "Name des abgespalteten Vereins";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["separationDate"][0] = "Datum der Trennung";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["separationDate"][1] = "Wann spaltete sich dieser Verein ab?";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["after_legend"] = "Nachfolger-Vereine";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["addAfter"][0] = "Nachfolger-Vereine hinzufügen";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["addAfter"][1] = "Hier können Sie ein oder mehrere Nachfolger-Vereine hinzufügen.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["afterClubs"][0] = "Bitte Verein(e) auswählen:";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["afterClubs"][1] = "Bitte wählen Sie ein oder mehrere Vereine aus.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["afterCause"][0] = "Grund";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["afterCause"][1] = "Bitte wählen Sie Grund für die Nachfolge aus.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["info_legend"] = "Information";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["info"][0] = "Vereinsbeschreibung";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["info"][1] = "Geben Sie hier eine kurze Beschreibung des Vereins ein.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["image_legend"] = "Bild/Logo-Einstellungen";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["addImage"][0] = "Ein Bild/Logo hinzufügen";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["addImage"][1] = "Dem Ereignis ein Bild/Logo hinzufügen.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["singleSRC"][0] = "Quelldatei";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["singleSRC"][1] = "Bitte wählen Sie eine Datei aus der Dateiübersicht.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["alt"][0] = "Alternativer Text";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["alt"][1] = "Hier können Sie einen alternativen Text für das Bild eingeben (<i>alt</i>-Attribut)";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["title"][0] = "Titel";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["title"][1] = "Hier können Sie den Titel des Bildes eingeben (<i>title</i>-Attribut)";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["size"][0] = "Bildbreite und Bildhöhe";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["size"][1] = "Hier können Sie die Abmessungen des Bildes und den Skalierungsmodus festlegen.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["imagemargin"][0] = "Bildabstand";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["imagemargin"][1] = "Hier können Sie den oberen, rechten, unteren und linken Bildabstand eingeben.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["imageUrl"][0] = "Bildlink-Adresse";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["imageUrl"][1] = "Eine eigene Bildlink-Adresse überschreibt den Lightbox-Link, so daß das Bild nicht mehr in der Großansicht dargestellt werden kann.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["fullsize"][0] = "Großansicht/Neues Fenster";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["fullsize"][1] = "Großansicht des Bildes in einer Lightbox bzw. den Link in einem neuem Browserfenster öffnen.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["caption"][0] = "Bildunterschrift";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["caption"][1] = "Hier können Sie einen kurzen Text eingeben, der unterhalb des Bildes angezeigt wird.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["floating"][0] = "Bildausrichtung";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["floating"][1] = "Bitte legen Sie fest, wie das Bild ausgerichtet werden soll.";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["link_legend"] = "Weblink";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["url"][0] = "Homepage";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["url"][1] = "Homepage des Vereins";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["live_legend"] = "Lebensdaten";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["intern_legend"] = "Intern";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["intern"][0] = "Bemerkungen";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["intern"][1] = "Interne Bemerkungen";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["publish_legend"] = "Veröffentlichung";
-
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["published"][0] = "Verein veröffentlichen";
-$GLOBALS["TL_LANG"]["tl_vereinsregister"]["published"][1] = "Den Verein auf der Webseite anzeigen.";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['publish_legend'] = 'Veröffentlichung';
+$GLOBALS['TL_LANG']['tl_vereinsregister']['published'] = array('Verein veröffentlichen', 'Den Verein auf der Webseite anzeigen.');
 
 /**
  * Buttons für Operationen
  */
 
-$GLOBALS['TL_LANG']['tl_vereinsregister']['new'][0] = "Neuer Verein";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['new'][1] = "Neuen Verein anlegen";
-
-$GLOBALS['TL_LANG']['tl_vereinsregister']['edit'][0] = "Verein bearbeiten";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['edit'][1] = "Verein %s bearbeiten";
-
-$GLOBALS['TL_LANG']['tl_vereinsregister']['editChronik'][0] = "Vereinschronik bearbeiten";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['editChronik'][1] = "Chronik des Vereins %s bearbeiten";
-
-$GLOBALS['TL_LANG']['tl_vereinsregister']['editNamen'][0] = "Vereinsnamen bearbeiten";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['editNamen'][1] = "Namen des Vereins %s bearbeiten";
-
-$GLOBALS['TL_LANG']['tl_vereinsregister']['editChairman'][0] = "Vorsitzende bearbeiten";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['editChairman'][1] = "Vorsitzende des Vereins %s bearbeiten";
-
-$GLOBALS['TL_LANG']['tl_vereinsregister']['editMembers'][0] = "Mitglieder bearbeiten";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['editMembers'][1] = "Mitglieder des Vereins %s bearbeiten";
-
-$GLOBALS['TL_LANG']['tl_vereinsregister']['copy'][0] = "Verein kopieren";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['copy'][1] = "Verein %s kopieren";
-
-$GLOBALS['TL_LANG']['tl_vereinsregister']['delete'][0] = "Verein löschen";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['delete'][1] = "Verein %s löschen";
-                                                          
-$GLOBALS['TL_LANG']['tl_vereinsregister']['toggle'][0] = "Verein aktivieren/deaktivieren";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['toggle'][1] = "Verein %s aktivieren/deaktivieren";
-
-$GLOBALS['TL_LANG']['tl_vereinsregister']['show'][0] = "Vereinsdetails anzeigen";
-$GLOBALS['TL_LANG']['tl_vereinsregister']['show'][1] = "Details des Vereins %s anzeigen";
+$GLOBALS['TL_LANG']['tl_vereinsregister']['new'] = array('Neuer Verein', 'Neuen Verein anlegen');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['edit'] = array('Verein bearbeiten', 'Verein %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['editChronik'] = array('Vereinschronik bearbeiten', 'Chronik des Vereins %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['editNamen'] = array('Vereinsnamen bearbeiten', 'Namen des Vereins %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['editChairman'] = array('Vorsitzende bearbeiten', 'Vorsitzende des Vereins %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['editMembers'] = array('Mitglieder bearbeiten', 'Mitglieder des Vereins %s bearbeiten');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['copy'] = array('Verein kopieren', 'Verein %s kopieren');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['delete'] = array('Verein löschen', 'Verein %s löschen');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['toggle'] = array('Verein aktivieren/deaktivieren', 'Verein %s aktivieren/deaktivieren');
+$GLOBALS['TL_LANG']['tl_vereinsregister']['show'] = array('Vereinsdetails anzeigen', 'Details des Vereins %s anzeigen');